# Scope reusing & smaller scopes