@nathanielc 2016-04-29
@yosiat Thanks! I was hoping this would also get done. At a quick glance it looks good to go once #491 gets merged.