@yosiat 2016-04-29
@nathanielc I think we need to define how it will be used.
I currently stuck in StatefulExpression but it might be more clean and ok to be it in it's own struct that StatefulExpression will own?