@nathanielc 2016-04-29
>I currently stuck in StatefulExpression but it might be more clean and ok to be it in it's own struct that StatefulExpression will own?

I think having it on its own would be good.