@yosiat 2016-04-29
@nathanielc I will fix that tomorrow and this PR will be ready to merge; I am going to sleep 💤