@nathanielc 2016-05-02
@yosiat Looks good. Can we also update the `eval.go` usage to use the scope pool? Basically any place that use `mergeFieldsAndTags` should be updated and then I think we can remove the `mergeFieldsAndTags` function.