@yosiat 2016-05-06
@nathanielc It's a bit problematic there because we have multiple expressions, we can create at startup "combined reference variables" set and use it, what do you think?