@nathanielc 2016-05-06
@yosiat I don't believe the scope is shared across the expressions, so they should all be able to operate independently.