@yosiat 2016-05-06
@nathanielc Looking at the code, looks like they use the same "vars", and after evaluation you are updating the scope..

It looks like the second expression is using the result of the first expression, etc.