@nathanielc 2016-05-06
@yosiat Yes, the `scope` cascades across expressions but the `state` is independent, and the set of referenced values is constant as well.

You should be able to use `ScopePool.Get` and check if the next expression even wants the result of the previous expression.  In other words create a scope for each expression and then only cascade needed values. The currently implementation is lazy and just reuses the same scope.