@yosiat 2016-05-06
umpf, so why not simply on upstart get the reference variables used in all scopes and merge them to one set? so the initial scope will be the merged? in term of performance your approach will win, but this approach will be much simpler.