@nathanielc 2016-05-06
Go with the simpler approach :) The biggest win is to remove allocations for each evaluation of the expression. Other than that lets keep it simple.