@yosiat 2016-05-06
What do you think about creating "MergedScopePool" which will accept list of nodes and and create scope pool for them?