@nathanielc 2016-05-06
I would rather not create another type. 

What if the `NewScope` method took a list of names `[]string` instead of a node. Then you could easily create ScopePools for difference use cases. You would need to expose a way to get the referencedVariables still. I think adding a method `ReferencedVariables` to the `StatefulExpression` would work as it has a reference to the node already. Then remove the `ScopePool` method from the `StatefulExpression`, in other words let the caller use the pool but don't require it. Then in `eval.go` on upstart you would call `ReferencedVariable` for each expression and merge them and create a single ScopePool for all the expressions. Does that makes sense? Basically instead of pushing the merge logic down into the `stateful` package leave it in the `eval.go` code since its the code that needs it.