@yosiat 2016-05-06
@nathanielc I will make "findReferenceVariables" a function on top of Node and will put under "tick/reflect.go"