@yosiat 2016-05-06
@nathanielc after thinking a bit, if the StatefulExpression won't own the scope pool, it means that:

- alert: will need to contain a symmetric array for holding the scope pools
- EvalPredicate: will need to accept the scope pool

Are we ok with those changes?