@yosiat 2016-05-06
@nathanielc This is getting some what tedious, In where, alert and eval we have to "maintain" symmetric array for holding the scope pools. I think it might hear readability.