@nathanielc 2016-05-06
@yosiat Then maybe you can supply the ScopePool to the StatefulExpression and it does the `maintaining`? But it still allows for the use case of the pool to be different.