@yosiat 2016-05-06
@nathanielc I thought about it, but then it will be too much weird, because they will some dependency.

In order to create ScopePool you will need reference variables, which you get by calling "ReferenceVariables" on StatefulExpression.