@nathanielc 2016-05-06
@yosiat OK, you are more in the details at this point. Do what makes sense...