@yosiat 2016-05-06
@nathanielc What do you think? I haven't integrated it in eval.go yet..