@nathanielc 2016-05-06
@yosiat 

>In order to create ScopePool you will need reference variables, which you get by calling "ReferenceVariables" on StatefulExpression.

This is essentially the only difference between alert/where and eval, is how the scope is created. It is used in the same manner. So either have a little duplicate code in alert and where or add helper functions to the `stateful` package. Something like `CreateScopePoolForExpressions` and it takes a list of expressions. Then eval passes in a list and where/alert pass in a list of one.