@yosiat 2016-05-06
@nathanielc do we want "CreateScopePoolForExpressions" to replace "NewScope" or to be helper method in stateful/scope_pool.go? (I think the latter)