@nathanielc 2016-05-06
@yosiat  The later, if something wants to create a scope in yet another different way it is still possible if we leave NewScope as is. Also that is why I called it `Create...` because there is implied logic happening not simply instantiating a new object.