@yosiat 2016-05-07
@nathanielc I started implementing the `CreateScopePoolForExpressions` in eval.go, and it looks like we have another problem.
We need `ReferenceVariables` for `fillScopePool (in expr.go)` and in eval.go we will need to have another list for it.

I think instead of `CreateScopePoolForExpressions` we will create: `FindReferenceVariablesForNodes` and remove it from Expression.

So the usage will be something like:
```go
var node tick.Node // let's say we have node
refVariables := stateful.FindReferenceVariablesForNodes(node)
stateful.NewScopePool(refVariables)
```

and in eval.go, it will something like this:
```go
refVariables := stateful.FindReferenceVariablesForNodes(n.Expressions...)
stateful.NewScopePool(refVariables)
```