@yosiat 2016-05-07
@nathanielc It's finished, squashed & updated benchmarks.