@nathanielc 2016-05-07
@yosiat :+1:  I'll merge on green. Since nothing user facing changes no need to write a CHANGELOG.